+++ HISTORY.rst
@@ -7,6 +7,8 @@ History
 - Add progress bar colors: blue while running, green on completion, red on failure (nens/rana-qgis-plugin#294)
 - Enforce .json suffix when saving importer presets to file (nens/threedi-schematisation-editor#493)
 - Validate settings files when importing via processing algorithm (nens/threedi-schematisation-editor#449)
+- Skip integration (with warning) when object matches to multiple conduits (nens/threedi-schematisation-editor#409)
+- Fix structures outside conduit raw bounding box not being matched despite being within snapping distance
 
 
 2.4.10 (2026-05-06)